Use previous status information
When checking status multiple times while using queueing features, use
the information provided by the json file to repeat the status of the
test.

While checking status during queueing operations, if a test is skipped
due to dependencies, verify that really is the case and alter the
status message if it is not. This was necessary when we (correctly)
began running the postProcess checks in reverse (reverse dependencies).

Refs #9000