output: remove confusing templated inheritance from Output class heirarchy
Previously, all subclasses of AdvancedOutput needed to specifiy certain
"output types" and an error was thrown if they weren't specified by
subclasses.  But now since there are subclasses that don't use any
special sub "output types", it is okay to not throw an error when this
is the case.  So We remove the error check.

Previously, the console output timestep information even if the object's
execute_on flags were set to not run on TIMESTEP_BEGIN. This should not
have been happening because the console was coded to only print timestep
information on TIMESTEP_BEGIN.  Which means Console::output was being
called during EXEC_TIMESTEP_BEGIN despite its flags - bad.  This
behavior was fixed, but we might want to provide some sort of legacy
support if other apps/users were relying on this in any significant way.

First step in addressing #9176