PorousFlowDarcyBase now includes counters that record, for each node in each element, the number of times it is an upwind node and the number of times it is a downwind node.  The counters are incremented every nonlinear iteration.  From this info the maximum number of upwind-downwind swaps in each element can be computed.  If, for any timestep, the number of swaps exceeds a user-input quantity, then the upwinding scheme is changed for that element at that timestep.
The point of changing the upwinding scheme is to improve
convergence, since with full upwinding if the upwind direction
is changing every nonlinear iteration the Jacobian will be
poor and the nonlinear convergence will be poor.

Currently there are two alternate schemes offered.
(1) "quick" in which the nodal mobility is used, just like
full upwinding, but no mass preservation is gauranteed.
(2) "harmonic" in which the element is assigned the harmonic
mean of the nodal mobilities.  This conserves fluid mass
but is typically slower than "quick".

In all these upwinding schemes, a crucial property is that
no fluid should be withdrawn from a node where the mobility
is zero.

Fixes #8901