Handle RunApp _only_ tests differently
When a test is strictly a RunApp test (like a python test) allow RunApp to set a
success status. Continue to set a failed statuses on failure events.

This is neccessary due to our Exodiff testers (and CSVDIFF and possibly others...)
having a success status being set when in fact, the test was not entirely finished
(see processResults in Exodiff --> FileTester --> RunApp). When that happens,
sometimes a prereq thread that was waiting for this test to _PASS_, launches. Thus
calling tester.prepare(). Which _deletes_ the output file to start anew, causing
the original Exodiff test to fail with "file not found errors" just as it started
to run an exodiff command. Fun find.

Refs #9280