Adding dependency resolution on control data
When getControlData is called we store the dependency on control data in
the RELAP7Control object. Note we cannot access the object yet, becuase
it may not exists. Further more, when we declareControlData(), we note
which RELAP7Control object built it. So when all objects are build we
can go over all RELAP7Controls and find out which control data they
depend on. From that, we get which control object built it and that is
the dependency we set for MOOSE to execute the controls in the right
order.

Refs idaholab/relap-7#437