Increase default max_its ref #6873
In our testing of this code on larger models, we have run into a couple of
cases where the residual started out really high (~1e100), and after 200
iterations, it was still making good progress, but it hadn't quite converged
just because of the huge initial residual. 300 seems to be sufficient to
converge every case I've tried it on now.