Avoid redundant MooseVariableScalar::insert()
This should fix the parallel scaling issue discussed in #9576.

This is a pretty deep change, so in theory there might be side
effects... but I can't picture anything being broken after this change
unless it was already broken-with-a-race-condition beforehand, e.g. if
different processors computed different values for the same scalar
variable.