Limit 1d_dg_block_restrict test to 12 procs
I verified that it works fine with MUMPS LU, so the convergence
failures we see at around 15 processors are just preconditioning
or iterative linear solver issues.  However, I don't see any *other*
MOOSE tests using MUMPS, and I don't want to be the one to introduce
that dependency, and neither of my other go-to preconditioner options
(BoomerAMG, ASM+ILU levels cranked up) worked readily, so we'll just
leave this as a low-processor-count test for now.