Determine restart status via mesh file extension
This is an ugly hack, but it should be correct for our purposes.
The only time we need to avoid repartitioning is when we read a
parallel restart.  The only time we have a parallel restart is when we
wrote in Checkpoint format.  Conversely, the only time we should ever
be reading Checkpoint format is to do an efficient restart; that's not
a stable file format so people shouldn't be archiving it.