Update for libMesh::MeshBase::is_serial_on_zero()
We'll need that method shortly: we need to be able to test whether a
displaced mesh has been automatically serialized to zero (e.g. for
Exodus output), so that we can keep the reference mesh up to date in
all respects, so that we can handle displacements of all nodes
properly as per #9626.