Disable libHilbert numbering on all restarts
At the moment, MOOSE can't do N-to-M restarts with .rd files, which
means MOOSE can't do N-to-M restarts.

If we can't do N-to-M restarts, there's no point in using libHilbert,
which adds a little CPU and a little more communication cost solely to
support N-to-M restarts.

Most importantly, we don't dare use libHilbert *inconsistently* -
trying to select libHilbert ordering based on the state of the
Adaptivity object was causing solution scrambling when user objects
manually adjusted adaptivity settings mid-simulation.

This fixes the second problem in #9652 for me, but probably shouldn't
close that issue because it doesn't fix the first problem and I'm not
sure if the initialSetup() workaround in his ToggleMeshAdaptivity test
object is acceptable to @jessecarterMOOSE as a long-term solution.