Allow components to setup mesh
All components must be able to do changes to the mesh.  Even
components that do not have mesh or generate one, must have a way to
do changes. For example, a boundary condition will want to generate a
side set which can then be used by users to add their additional objects
on. This is a great help to UX, because boundary IDs can change when
components change. Names are more persistent.

Refs idaholab/relap-7#19688