Use BoundingBox, not MeshTools::BoundingBox
This gets rid of several deprecated libMesh API calls.

In particular, it gets rid of a *buggy* API call, where
MultiAppTransfer was trying to get a local bounding box via misuse of
processor_bounding_box() and was inadvertently getting a global
bounding box instead.  This was preventing us from merging
https://github.com/libMesh/libmesh/pull/1303 and it resembles other
buggy calls which have in the past caused #1500 tests to fail.