Run RandomIC based tests in serial on one thread
If I'm correctly understanding #5567 and the code, we still don't
expect that InitialCondition to give us reproduceable results with
more than one MPI rank or more than one thread.

There are lots of other RandomIC based tests which I'm leaving enabled
in parallel, at least for now, because they passed for me with -p 2.
The ones I glanced at are *expected* to succeed in parallel, because
they use eigensolves or only test jacobians of linear kernels or are
in some other way independent of initial conditions.  So, I'm not
going to force any RandomIC test to serial until I actually *see* it
fail in parallel.