AnnularMesh added
This allows users to create 2D annular or disc meshes.  The meshes are simple: they are created by drawing a number of radial lines and a number of concentric rings, and the mesh is the quads that are thus described (along with triangles, in the case of a disc mesh).  Appropriate boundary sidesets are added.

Fixes #8432