Change unknown variable name in 1D advection tests to 'c'.
This is meant to stand for e.g. "concentration". IMO it was very confusing
having a parameter 'u' which represented the advective velocity and an
unknown with the same name.

This requires regolding the test files; I verified that the only
difference between the old and new files was the variable names.  This
is follow on work to #10138.