VanGenuchten capillary pressure with nonzero derivative.
The van Genuchten capillary pressure may now be modified so the derivative dS/dP<0 for S->1.  This is achieved by setting a scaling parameter called s_scale < 1.

It will be unusual for users to set s_scale < 1, but i've found that in occassional models this helps with convergence: when the system is (almost) fully water saturated and the two nonlinear variables are porepressures.

Fixes #10181