parser: make filepaths in input relative to input file dir
All input parameters that are typed FileName, FileNameNoExtension, or
MeshFileName (and vectors of these types) are treated as relative to the
directly the input file itself resides in.  The parser now prepends a
file path prefix to enforce this.

This will help with benchmarking and also pave the way for sandboxing of
tests/benchmarks in the future if we decide it is worth while.

ref #9834