testharness: only run tests in CWD subtree
Instead of the run_tests script running all tests in its dir subtree
using a specific named binary, modify the test harness to just run tests
in the current working directory's subtree.  The binary is discovered
walking of the parent directory hierarchy looking for a "testroot" file.
This file contains the name of the binary to be used.  "run_tests"
scripts now can all be identical with no app-specific content.

Also provide a "run_tests" script that users can add to their path that
will always run tests in the CWD for *any* moose app/binary - it just
hunts for a "testroot" file.

fixes #10209