Add better LaplacianJumpIndicator tests.
These new tests are based on the Biharmonic equation (and its
transient variant) so the Laplacian jump indicator is appropriate for
estimating the error.

In the process of developing these tests, it was discovered that MOOSE
is not currently imposing Dirichlet BCs for Hermite elements
correctly. Specifically, MOOSE constrains only the *value* degrees of
freedom when Dirichlet BCs are imposed, but setting BCs for Hermite
elements involves the gradient degrees of freedom as well, according
to @roystgnr. The current workaround is to use a penalty approach for
enforcing the boundary conditions, but this comes with the obvious
drawbacks of dramatically increasing the initial residual and changing
the conditioning of the system.

Refs #2190.