Nodes are displaced via MOOSE objects
In order to accomodate adaptivity, we have to start displacing nodes
using the MOOSE system.  There are several things to be addressed:

1. With adaptivity, we can no longer assume contiguous node numbering,
so the way we displace nodes is not any good anymore.

2. New nodes produced by adaptivity have to be displaced, so we need to
do so after the mesh changes.

3. The x-componnet of gravity needs to be computed _after_ nodes after
displaced, becuase this computation relies on the proper node position
in the physical space.

Refs idaholab/relap-7#487