parser: more robust handling of number syntax errors
stoi and stod functions are quite permissive and can ignore part of the
input string leaving it not entirely parsed.  Sometimes this could
result in invalid number strings being parsed as valid.  For example
"3ee10" is parsed as the value "3".  This change fixes that and requires
sto[i/d] to parse the entire string.  Any corresponding errors are not
added to the _errmsg collection variable and printed together with
other errors all at once instead of one at a time.

fixes #10310