Adding a helper script to report job status
Adding a script to facilitate the need for checking if all launched
PBS jobs have finished in the queue. Previously, folks (civet) would
run run_tests repeatedly, and grep for '0 pending jobs' in order to
determine when the process was complete. This was grossly inefficient.

When the script detects all jobs are ready, it exits with code 0.

This script also adjusts the status to 'waiting', so that when
run_tests _is_ executed, the TestHarness does not call qstat a second
time, since we already know the test is ready to be processed.

Closes #10326