Fix how --recover and syntax tests behave
Do not allow syntax checking tests to run with a recover
test.

While solving the above, found an issue with the way statuses
worked for discovered parser errors in tests. Created a unit
test for this. Also instead of 'skipping' a failed parsed
test file, we will now fail that test.

Including a way to adjust the column length based on an
influential environment variable: 'MOOSE_TERM_COLS'.

Closes #10400