Adding scalar zero into ScalarCoupleable
Scalar zero is equivalent of zero in normal coupling, but for scalar
variables.  This version is needed, so that when people use it, they get
properly sized zero for their scalar variable. The _zero would be sized
to the number of q-points, while we need number of components.  However,
to make things easier we always resize to order 43, becuase we cannot
have higer number of component for scalar variables (libMesh
limitation).

Refs #10630