Do not try to solve multiapps when the execution time is beyond their end time
If a subapp restricted its end time via end_time parameter, the master
app would try to solve it beyond this time point. The subapp would first
generate dt = 0 which would fail (think of du/dt terms), then it would
try to take smaller steps, but those would also failed and the whole
setup would not finish.

With this change we are not trying to solve the subapps whose end time
is smaller then the current global execution time.

Closes #4720