Allow MetaActions to add Actions with multiple tasks
Prior to this patch MetaActions could only add Actions
to satisfy a single task because we didn't know if conflicting
syntax could get in the way. This is often NOT the case, so
we can inspect the associated syntax and allow the newly
created Action to satisfy all tasks if said Action is
associated with zero or one syntax blocks.

closes #10619