Adding API for setting output dimension
When using lower dimensional elements in higher dimensional space, an
application might require to force the dimension of the Exodus output.
If this is supposed to happen via an action, an API for this is needed.
The manipulation of input parameters is not possible, becuase the object
already exists.

Also, requiring users to enter an extra line into all their input files
is not friendly, becuase we know what the dimension always is, so we
just need to set it for them.

Refs #9205