Improve test robustness
These tests occasionally fail on Macs. I looked into it, and it appears that
the recursion is chewing up a lot of stack space with all of the adpativity that
these tests have. For some odd reason, this sometimes causes the Mac tests to just
run out of memory? Adding more procs makes these tests more stable.

Additionally, forcing the test to run in parallel uncovered a nasty bug
that would only be present with this test in a parallel run!

refs #7816