Change patch_update_strategy in some simple_contact tests.
This is required to keep these tests running warning-free with older
versions of PETSc (3.4.5). Another option would be to add
'allow_warnings=true' to the specs for these tests, but I think this
approach is more robust.

Without this change, on PETSc 3.4.5, the warning in question is given below:
```
Warning in PenetrationLocator. Penetration is not detected for one or
more slave nodes. This could be because those slave nodes simply do
not project to faces on the master surface. However, this could also
be because contact should be enforced on those nodes, but the faces
that they project to are outside the contact patch, which will give an
erroneous result. Use appropriate options for 'patch_size' and
'patch_update_strategy' in the Mesh block to avoid this issue. Setting
'patch_update_strategy=iteration' is recommended because it completely
avoids this potential issue. Also note that this warning is printed
only once, so a similar situation could occur multiple times during
the simulation but this warning is printed only at the first
occurrence.
```

Refs #10906.