Fix 'are_my_queued_jobs_finished'
The script was altering the json status in such a way that the
TestHarness was no longer checking for the exit code (by running
qstat). This was initially by design, as running qstat is a timely
process.

The following change, modifies the script so that -it- continues to
function quickly, but not so, for the TestHarness. This forces the
TestHarness to make that final slow call to qstat, so it now
understands the exit code properly.

Closes #10931