1D flow in 3D space
Changing 1D kernels to work in 3D non-displaced space.

Also, changing the way gravity is computed. The projection of the gravity
vector in the x-direction was precomputed by an aux kernel.  Now, that we
have the pipe direction computed in a material, there is no need for
this, we just need to bring in the global gravity vector and dot it with
the pipe direction.