Use local files in PeacockApp tests
These tests seem to fail randomly. This is most
likely due to there being a window when we set
the input file in the common dir, the mesh gets generated,
then we change directory to the test dir.
During this time other tests might be doing the same thing,
causing corruption in the mesh file, which generates VTK errors,
which causes crashes.

closes #11296