Add error checking args to SubProblem::getVariable().
The additional arguments have sensible default values so the change
should be backwards compatible. I also went through the framework,
modules and apps to figure out which flags were appropriate to
pass in various getVariable calls. The following classes were updated
in modules:

* CrackFrontDefinition
* CrackFrontData
* RichardsMultiphaseProblem
* LevelSetBiMaterialBase
* LevelSetCutUserObject

Refs #11192.

A note on default values for virtual functions:

According to the standard, default values are determined by the
_static_ type used to make the call to the function. Since
getVariable() is frequently called through an FEProblem reference, its
getVariable() override must provide the defaults. A good rule of thumb
is: "Make default arguments the same or absent when overriding a
virtual function."