New default args should go at end of parameter list.
This is the most backwards-compatible approach, although I did not
come across any apps that were broken by the other order.  Also changed
the default values to "ANY" for both, as this is also the most backwards
compatible behavior.

Use the new constructor for MultiAppDTKUserObjectTransfer, apparently
I forgot to do that one before...