Postprocessor names can be used directly in the control system
We added an action that will create an instance of
CopyPostprocessorValueControl every time a postprocessor is added. This
control object takes a postprocessor value and copies it into a control
data structure. This is so that it can be used inside the control
system.  It is important to do this, so that the proper dependency is
maintained (i.e. values are copied before they are actually used).

This should increase user friendliness of the control system. Also
reduces the size of input files.

Refs idaholab/relap-7#437