Fix AnalyzeJacobian race conditions
A race condition exists when multiple AnalyzeJacobian using the
same input file runs simultaneously.

This change detects these issues. Also added a unittest for this
case scenario.

During the creation of this PR and the unittests, a typo in fact,
a new issue was discovered #11454 which will require a different
approach.

Closes #11433