Removing physical mesh
This variable is no longer needed, becuase our mesh is the physical
mesh.  Cleaning up left overs from earlier change that removed the need
for displaced problem.

Also marking clone method in RELAP7Mesh as problem, becuase it will be
removed in upcoming MOOSE update.

The gold file was changed becuase the original had dimension = 2, now
the dimension is 3 (as it is supposed to be).

Refs idaholab/relap-7#19688