Adding ADDITIONAL_DEPEND_LIBS to library targets
The ADDITIONAL_DEPEND_LIBS variable is useful for creating extra external dependencies.
Right now this flag is only part of the executable, but it really needs to be passed to the
linker for library creation as well or you end up with libraries that don't know about
their dependencies. We already have EXTERNAL_FLAGS in the library and executable creation
but we need the ADDTIONAL_DEPEND_LIBS variable added to the dependency list as well in
case downstream Makefiles are building dependency libs on demand.

refs #10574