Add test verifying Nemesis variable hiding actually works.
The original version of this test was not sufficient to exercise an
indexing bug in libmesh, the new version of the test is more complete
in this regard. The new test has multiple nonlinear and auxiliary
variables of both nodal and elemental types. I have also tested this
on various combinations of hidden/unhidden vars and aux vars and the
variable hiding code seems to work for all of them.

In MOOSE, what currently happens is that only nodal variables are
written to Nemesis files, so that CONSTANT, MONOMIAL auxes don't show
up in the output at all. This is due to the fact that we automatically
hide all variables, except those in getNodalVariableOutput(), in
Nemesis.C.

There remain two versions of this test which are specific to different
compiler/architecture combos, since Parmetis partitioning is still
known to be architecture dependent (due to different random number
generators being used on different systems).  This test now has its
own input file rather than piggybacking off others.

Refs #11469.