Input porosity to PorousFlowPorosity classes is now constant monomial.
Only the [0] value of this is used, not [_qp], because it is a constant monomial.  This is advantageous because it means that we get the same values, and can be assured of non-erroneous indexing, regardless of whether this is a nodal Material (computing values of porosity at the nodes of the element) or an ordinary Material.

Fixes #11620