Adding coupled convective heat flux BC
The CoupledConvectiveFlux was deprecated and replaced by more general
CoupledConvectiveHeatFluxBC.  The latter one allows the convective heat
transfer coefficient to be a spatially varying field. It can also
accomodate multi-app needs, where field are computed by an external code
and transferred into auxiliary variables and then used.

Users can also use vector coupling of variables to specify multiple
phases for the heat flux.  When doing so, users are required to provide
`alpha` parameter which is the volume fraction of each phase.

Closes #11631