Changed hard-coded usages of RELAP7::gravity_const to the correct value
Tests using TRACE closures and a zero gravity vector would diff because
TRACE closures were the main culprits of using RELAP7::gravity_const.
Ideally, imposing a perpindicular gravity vector with magnitude equal
to RELAP7::gravity_const would be all that is necessary to fix the tests
without re-golding; however, horizontal closures are not implemented for
all TRACE closures, so the gravity needed to be along the pipe direction.
As a result, the now-present gravity terms required these tests to be
re-golded.

Refs idaholab/relap-7#555