Added enums for gravity angle
It is believed that the binary "is horizontal?" classification of
pipe orientation with respect to gravity is insufficient. For instance,
there is the zero gravity case, which is neither horizontal nor
vertical, and also, there may be closures that impose that the flow
channel is completely horizontal/vertical, not just mostly horizontal/vertical.

Refs idaholab/relap-7#555