fix incorrect line number calculation in hit
When the consecutive string literal syntax was added to hit, line number
accounting became broken.  There was a case where we "rewinded" back to
the end of a prior token across previously ignored whitespace that may
have included newlines.  Previously, this rewinding did not also rewind
the line count accordingly.  Fix this and add a test to prevent
regressions.

Fixes #11622