Final draft of the concentric circle mesh C file (#11656)
This is to create a 1/4 typical LWR fuel pin. Users need to specify
the fuel pin radius, the inner cladding radius, and the outer cladding
radius. It automatically adjusts radiuses to keep the actual target
volumes. In addition, the numbers of intervals for the fuel pin and
the moderator part must be provided. The number of sectors can be
modified to make the mesh either coarser or refiner.

Closes #11656.