Throw a better error when case is incorrect in registration
If an object is used in the input file and it's registered incorrectly,
the parser won't even be able to build that object and would normally
just tell the user it wasn't registered. This is not a useful error because
the object might indeed be registered, but just under a nonsense label.
To combat this case we'll check to make sure that the label is OK right
before we are going to die anyway due to an unregistered error.

Note: There's not really a good way to test for this condition, it would
require a separate poorly registered object somewhere, but since we report
all incorrectly registered objects, that entire application would only
serve to report this error. It's still worth reporting because it's a
relatively easy mistake to make in a new app (e.g. case difference).

closes #11644