Split Grain Test Cleanup
The splitting grain test was added here: 119a7cc4d33cf83db0543e3b0701ee0515fab776
once the necessary feature was implemented. However, the original test file
was never removed. Cleaning that up.

Switching the splitting grain test to use "bt" instead of "jp" because
bt does a better job at distributing grains with low numbers of ops. In
this test it fails to seperate two grains in the IC (they overlap on halos),
which causes the unique id of one grain to change on the first time step
(treated as a single grain).

refs #7712