Do not process the same finished job twice
When a long running alert occurs, it was possible for the job to finish before
the alert was processed in the status pool. This caused job_count discrepancies
(and thus the exception) when deducting the amount of finished jobs the status
pool was processing.

This PR eliminates that issue, as well as using a set to determine if all jobs
submitted have been processed.

Closes #11932