Removing pure abstract requirements
The base class  has all methods implemented with an error message, so
that when the method is needed, users get an error. We do not want pure
abstact requirements, because not all methods has to be implemented in
order to run a problem successfully. This was users will have much
easier time add new fluid properties.

Refs #12050