Added EXEC_INITIAL to boundary flux user objects
If residuals are computed on INITIAL, then boundary fluxes
should execute on INITIAL. The only thing that happens when
boundary fluxes execute is the initialize() function, which
initializes the indices corresponding to cached values. If
this is not executed before the first call to getFlux(), then
a segmentation fault may occur due to unitialized cached value
indices happening to get real values such as 0.