Added ghost boundary flux for mass flow rate and temperature for 1-phase
There is already an implementation for the mass flow rate and temperature
boundary flux for 1-phase, but it does not use a ghost cell. This commit
adds the ghost cell implementation, but it not plugged into the boundary
component at this point. Later, the user could be given the option of
which to use. The ghost cell implementation was made to compare to the
non-ghost version to ensure that the resulting solution is still
satisfactory. The reason for checking this is that for 2-phase, it would
be much simpler to implement the ghost cell version.

Refs idaholab/relap-7#484