Allow tests to specify trim threshold
Some tests need to scrutinize output within the frame which was trimmed.

This feature will probably be abused in the fact that no one will use an
actual number. -1 effectively turns off trimming (equivalent to --no-trimmed-output).

Added a unittest which _does_ set a real number. A real small one.
Which triggers the code that injects "Output trimmed". This is what the
unittest will search for.

Closes #12167