Fix diffing HPC modules tests closes #12209
I think the best way to fix the issue with sorting of sampler entries is to
just incline the line for this test. I don't think it's really that big of
a problem that they might not be ordered consistently in cases like this,
at least not a big enough deal to warrant the messy code that would be involved
in fixing the issue.

For the global strain test, I added a custom comparison file to avoid applying
loose tolerances across the board.

Also ref #7387