Adding a "try-harder" loop to the GrainTracker
This loop is intended to look at the previous state to match up
new grains in an attempt to properly classify them as split grains
as opposed to a new grain (nucleated). There's more to do here, but
this is a reasonable start that will work for a number of cases.

What's missing is a good way to test this. I'll try to work something
up, but it's a very difficult case to make happen.

closes #12236